test: Improve x/evm genesis tests (#73)
* test: Improve x/evm genesis tests

Previously, the InitGenesis function did not have complete unit
coverage, did not check for panic values, and shared a test suite with
the handler.  This created dependencies on setup for unrelated tests and
made it difficult to add, change, or refactor genesis behavior with
confidence.

These changes use a fixture based approach to testing InitGenesis which
improves the ability to define more complex test cases and run each sub
test in isolation.  Once global state is removed from app setup, this
enables an easier transition to parallel tests.

This comes from the ability to setup the application, define a function
for each test that is provided a context and app, then generates a
fixture with required dependencies and expectations.  This allows
context to be shared across state setup, expectations, and panic values.

In addition, all InitGenesis behavior was checked in TDD fashion and
tests were added where coverage was lacking before.

* refactor: ExpectFunc default should be nil

When defining a test fixture with no expectations, the test fixture
should use a nil function for expectFunc, instead of an empty function
which serves no purpose.

This also ensures we check for nil before calling the set function in
order to prevent panics.

* refactor: Test Cases names should be capitalized

When reading, parsing test names from file, or reading verbose test
output, a capital test name is much clearer to read and comprehend than
an uncapitalized one. We mimic git commit message subjects in this
sense.

* fix: Fix incorrect capitalization of panic value

I screwed this up when renaming test names, correct expected value.